requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
slowapi==0.1.8
limits==3.6.0
psycopg2-binary==2.9.9  # Для PostgreSQL
asyncpg==0.29.0  # Для асинхронной работы с PostgreSQL
python-dotenv==1.0.0